refactoring - moving ruby.rb - all writer logic into their own separate classes by type - lot easier to manage
