+++ lib/ffi_generator/generators/from_c_generator/by_value_type.rb
@@ -10,6 +10,14 @@ module FFIGenerate
         @inner_type.name
       end
 
+      def ruby_name
+        @inner_type.ruby_name
+      end
+
+      def ruby_ffi_type
+        "#{@inner_type.ruby_ffi_type}.by_value"
+      end
+
     end
   end
 end
