+++ lib/ffi_generator/generators/from_c_generator/generator.rb
@@ -24,13 +24,15 @@ module FFIGenerate
                 :prefixes,  # TODO: deprecated
                 :suffixes   # TODO: deprecated
 
+    attr_accessor :writer
+
     def initialize(options = {})
       @module_name   = options[:module_name] || fail("No module name given.")
       @headers       = options[:headers] || fail("No headers given.")
       @cflags        = options.fetch(:cflags, [])
       @ffi_lib       = options[:ffi_lib]
       @ffi_lib_flags = options[:ffi_lib_flags]
-      @ffi_lib_path_env = "#{@module_name.upcase}_PATH".gsub(/[\W]+/, "")
+      @ffi_lib_path_env = ffi_lib_path_env
       @blocking      = options.fetch(:blocking, [])
       @output        = options.fetch(:output, $stdout)
       @translation_unit = nil
@@ -40,7 +42,13 @@ module FFIGenerate
       @prefixes      = options.fetch(:prefixes, []) # TODO: deprecated
       @suffixes      = options.fetch(:suffixes, []) # TODO: deprecated
 
+      @writer = Writer.new "  ", "# "
       auto_include_c_paths
+    end
+
+    def ffi_lib_path_env
+      flat_name = @module_name.upcase.gsub("::", "_")
+      "#{flat_name}_PATH".gsub(/[\W]+/, "")
     end
 
     def generate
@@ -82,10 +90,61 @@ module FFIGenerate
       end
     end
 
+    def generate_rb
+      modules = @module_name.split("::")
+
+      writer.puts "# [generated by ffi_generator] #", ""
+      writer.puts "require 'ffi'", ""
+
+      write_modules_block(modules) do
+        write_header
+        declarations.each(&:write_ruby)
+      end
+
+      writer.output
+    end
+
+    def write_modules_block(modules, &block)
+      write_one_module_block(modules, 0, modules.length, &block)
+    end
+
+    def write_one_module_block(modules, index, length, &block)
+      writer.puts "module #{modules[index]}"
+      writer.indent do
+        if index < length - 1
+          write_one_module_block(modules, index + 1, length, &block)
+        else
+          yield
+        end
+      end
+      writer.puts "end"
+    end
+
+    def write_header
+      writer.puts "", "extend FFI::Library", ""
+      writer.puts "ffi_lib_flags #{@ffi_lib_flags.map(&:inspect).join(', ')}" if @ffi_lib_flags
+
+      if @ffi_lib
+        msg = "ffi_lib ENV[\"#{@ffi_lib_path_env}\"] || #{@ffi_lib.inspect}"
+        puts "Setting lib path as: #{msg}"
+        writer.puts msg, ""
+      end
+
+      write_attach_function
+    end
+
+    def write_attach_function
+      writer.puts \
+        "def self.attach_function(name, *_)",
+        "  begin; super; rescue FFI::NotFoundError => e",
+        "    (class << self; self; end).class_eval { define_method(name) { |*_| raise e } }",
+        "  end",
+        "end",
+        ""
+    end
+
     def translation_unit
       return @translation_unit unless @translation_unit.nil?
-
-      args = @cflags
 
       index = Clang::Index.create
