+++ lib/ffi_generator/generators/from_c_generator/pointer_type.rb
@@ -13,6 +13,18 @@ module FFIGenerate
         @pointee_name
       end
 
+      def ruby_name
+        @pointee_name.to_ruby_downcase
+      end
+
+      def ruby_ffi_type
+        ":pointer"
+      end
+
+      def ruby_description
+        "FFI::Pointer(#{'*' * @depth}#{@pointee_name ? @pointee_name.to_ruby_classname : ''})"
+      end
+
     end
   end
 end
