+++ lib/ffi_generator/generators/from_c_generator/string_type.rb
@@ -6,6 +6,14 @@ module FFIGenerate
         Name.new(["string"])
       end
 
+      def ruby_name
+        "String"
+      end
+
+      def ruby_ffi_type
+        ":string"
+      end
+
     end
   end
 end
