+++ lib/ffi_generator/generators/from_c_generator/type.rb
@@ -2,6 +2,10 @@ module FFIGenerate
   class Generator
     class Type
 
+      def ruby_description
+        ruby_name
+      end
+
     end
   end
 end
