+++ lib/ffi_generator/generators/from_c_generator/unknown_type.rb
@@ -6,6 +6,14 @@ module FFIGenerate
         Name.new(["unknown"])
       end
 
+      def ruby_name
+        "error_unknown"
+      end
+
+      def ruby_ffi_type
+        ":char"
+      end
+
     end
   end
 end
